+++ changelog.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for introspecting domain types.
+  ([#380](https://github.com/hasura/ndc-postgres/pull/380))
+
+### Changed
+
+### Fixed
+
 ## [v0.5.1] - 2024-03-21
 
 ### Added